pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mymory"
version = "0.1.0"
description = "Governed memory vault substrate with wings, corridors, and scheduled hygiene. MyMories + MemPalace + Graphify hybrid."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Sean Uddin" },
]
keywords = [
    "memory",
    "obsidian",
    "knowledge-graph",
    "rag",
    "agentic",
    "offline",
    "sentence-transformers",
    "graphify",
    "mempalace",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Utilities",
]

dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "rich>=13.0",
    "python-frontmatter>=1.1",
    "sentence-transformers>=2.7",
    "numpy>=1.26",
    "pymupdf>=1.24",
    "python-docx>=1.1",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1",
    "beautifulsoup4>=4.12",
    "markdownify>=0.11",
    "pandas>=2.2",
    "striprtf>=0.0.26",
    "watchdog>=4.0",
]

[project.optional-dependencies]
lancedb = ["lancedb>=0.6"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "black>=24.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
mcp = [
    "mcp>=1.0",
]
all = [
    "mymory[lancedb,mcp]",
]

[project.urls]
Homepage = "https://github.com/uddin/mymory"
Documentation = "https://github.com/uddin/mymory/blob/main/README.md"
Repository = "https://github.com/uddin/mymory"
Issues = "https://github.com/uddin/mymory/issues"

[project.scripts]
mymory = "mymory.cli:main"

[tool.setuptools]
packages = [
    "mymory",
    "mymory.core",
    "mymory.parsers",
    "mymory.graphify",
    "mymory.mempalace",
    "mymory.layer2",
    "mymory.layer3",
    "mymory.scripts",
]

[tool.setuptools.package-data]
mymory = [
    "config/*.yaml",
    "config/*.yml",
    "plugin/.claude-plugin/*",
    "plugin/skills/**/*",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-select = ["I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra --strict-markers"

[tool.mypy]
python_version = "3.10"
strict = false
warn_unused_ignores = true
ignore_missing_imports = true
